test: fix put with hash test
We just changed ID/"id" to IDENTITY/"identity" to match the multicodec table and
avoid confusion with peer IDs, CIDs, etc.

Unfortunately, this breaks the `--hash=id` flag. Luckily, I'm pretty sure
nobody's actually using this. As putting a block with an identity hash is
useless. If they are users, we can go about adding in backwards compatibility
hacks later.